Split models and add type column
