+++ db/migrate/20170717150455_add_type_to_datafiles.rb
@@ -1,0 +1,5 @@
+class AddTypeToDatafiles < ActiveRecord::Migration[5.1]
+  def change
+    add_column :datafiles, :type, :string
+  end
+end
